backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
email-validator==2.2.0
SQLAlchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
openai==1.59.7
httpx==0.28.1
pdfplumber==0.11.5
pandas==2.2.3
openpyxl==3.1.5
gspread==6.1.4
google-auth==2.37.0
APScheduler==3.11.0
slowapi==0.1.9
structlog==24.4.0
tenacity==9.0.0
pytest==8.3.4
pytest-asyncio==0.25.0
aiosqlite==0.20.0
